Allow latex_escape to work in a Rails 2.2+ environment
Since Rails 2.2 also defines a #l alias, but for #localize, any use of #l
in non-rtex code was getting captured by rtex. Since the semantics of
localize and latex_escape are not the same, this led to unexpected behavior.

This patch causes latex_escape to pass work on to localize if (1) rtex
rendering is not being done and (2) if localize is defined (to allow
backwards compatibility with pre-Rails 2.2 applications).