Fix link generation error
When generating a link in StylesHelper#create_description_link, the
controller name needs to be prefixed with a / to prevent a routing
error when running the admin functional tests (the non-admin functional
tests are unaffected). However, in a non-test environment, a leading /
on the controller name generates a routing error in all situations.

The fix is to conditionally include the leading / on the controller
name depending on the environment.