Improved creation of new categories/awards/styles
When creating a new Category, the subform for a new associated Award is
automatically opened. Likewise, when creating a new Award, the subform
for a new associated Style is automatically opened. This is similar to
the behavior when creating a new Entrant and makes it a lot easier to
create new Categories, Awards, and Styles since the entire tree can be
created from the Categories form instead of having to bounce between
the separate forms.