Fix timezone offset calculations
Translations between local and UTC time were fubar, caused in part by
the unexpected usage of GMT offsets in TZInfo. TZInfo uses the Olson
timezone database, which uses POSIX rules for GMT offsets, and these
offsets are the reverse of conventional usage (i.e., specifying a
timezone of GMT-6 actually sets a timezone that is 6 hours EAST of GMT,
not 6 hours WEST).

Because of the potential for confusion regarding this issue, the
GMT-based timezones have been removed from the select list in the
competition data form. This should not be a problem, however, since
timezones specified as an offset from GMT are incapable of supporting
daylight saving time (because DST varies by region).